+++ utilities/backupable/backupable_db_test.cc
@@ -977,7 +977,7 @@ TEST_F(BackupableDBTest, FileCollision) {
 
     // If the db directory has been cleaned up, it is sensitive to file
     // collision.
-    DestroyDB(dbname_, options_);
+    ASSERT_OK(DestroyDB(dbname_, options_));
 
     // open with old backup
     OpenDBAndBackupEngine(false /* destroy_old_data */, false /* dummy */,
@@ -996,7 +996,7 @@ TEST_F(BackupableDBTest, FileCollision) {
     CloseDBAndBackupEngine();
 
     // delete old data
-    DestroyDB(dbname_, options_);
+    ASSERT_OK(DestroyDB(dbname_, options_));
   }
 }